fix: header 삭제 및 MainLayout 수정
